Add CSV Exporter for types
