+++ .gitignore
@@ -1,4 +1,5 @@
 **/node_modules
 **/out
 .env
-.DS_STORE
+**/.DS_STORE
+neo4j/
