+++ prototypes/analysis/Makefile
@@ -1,3 +1,6 @@
 run:
 	go build -o bin/analysis && bin/analysis
 
+build:
+	go build -o /bin/analysis
+
