--- prototypes/analysis/extract/extract_test.go
@@ -1,10 +1,0 @@
-package extract
-
-import (
-	"testing"
-)
-
-func TestGoExtract(t *testing.T) {
-	e := NewGoExtractor()
-	e.Extract("github.com/humanbeeng/lepo/prototypes/go-testdata")
-}
