+++ prototypes/analysis/go-types.csv
@@ -1,0 +1,13 @@
+name,qname,typeqname,underlying,code,doc
+NiceGreeter,github.com/humanbeeng/lepo/prototypes/go-testdata/current.NiceGreeter,`github.com/humanbeeng/lepo/prototypes/go-testdata/current.NiceGreeter`,`struct{Name string}`,type NiceGreeter struct {\n	Name string\n},
+FancyGreeter,github.com/humanbeeng/lepo/prototypes/go-testdata.FancyGreeter,`github.com/humanbeeng/lepo/prototypes/go-testdata.FancyGreeter`,`struct{}`,type FancyGreeter struct{},
+RudeGreeter,github.com/humanbeeng/lepo/prototypes/go-testdata.RudeGreeter,`github.com/humanbeeng/lepo/prototypes/go-testdata.RudeGreeter`,`struct{}`,type RudeGreeter struct{},
+Person,github.com/humanbeeng/lepo/prototypes/go-testdata.Person,`github.com/humanbeeng/lepo/prototypes/go-testdata.Person`,"`struct{Name string ""json:\""name\""""; Age int ""json:\""age\""""; TypesInfo go/types.Info; Router github.com/gofiber/fiber/v2.Route}`","// Comment above struct decl\ntype Person struct {\n	Name      string \`json:\""name\""\` // Inline comment\n	Age       int    \`json:\""age\""\`\n	TypesInfo types.Info\n	Router    fiber.Route\n}","Comment above struct decl
+"
+MedicalData,github.com/humanbeeng/lepo/prototypes/go-testdata.MedicalData,`github.com/humanbeeng/lepo/prototypes/go-testdata.MedicalData`,`struct{Bloodtype string}`,// First line\n// Second line\ntype MedicalData struct {\n	// Field doc comment\n	Bloodtype string // Field inline\n},"First line
+Second line
+"
+EducationData,github.com/humanbeeng/lepo/prototypes/go-testdata.EducationData,`github.com/humanbeeng/lepo/prototypes/go-testdata.EducationData`,`struct{Degree string; College string}`,/*\nBlock comment above struct\n*/\ntype EducationData struct {\n	Degree  string\n	College string\n},"Block comment above struct
+"
+FinancialData,github.com/humanbeeng/lepo/prototypes/go-testdata.FinancialData,`github.com/humanbeeng/lepo/prototypes/go-testdata.FinancialData`,`struct{Networth int; BankData struct{BankName string; IFSC int}}`,type FinancialData struct {\n	Networth int\n	BankData struct {\n		BankName string // Inner inline comment\n		IFSC     int\n	}\n},
+BankData,github.com/humanbeeng/lepo/prototypes/go-testdata.FinancialData.BankData,`struct{BankName string; IFSC int}`,`struct{BankName string; IFSC int}`,struct {\n	BankName string // Inner inline comment\n	IFSC     int\n},
